pyproject.toml
[tool.poetry]
name = "pipe-crud"
version = "0.1.0"
description = ""
authors = ["socar-humphrey <humphrey@socar.kr>"]
license = "UNLICENSED"

[tool.poetry.dependencies]
python = ">=3.8,<3.9"
fastapi = "^0.68.0"
dependency-injector = "^4.35.2"
databases = "^0.4.3"
uvicorn = "^0.14.0"
ulid-py = "^1.1.0"

[tool.poetry.dev-dependencies]
pytest = "^6.2.4"
black = "^21.7b0"
sqlfluff = "^0.6.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
